Audio Unit (AU) plugin (#642)
* Implemeting audio call back method

* AU plugin running in Reaper and Logic 9

* Creating NSView programatically

* Renaming view classes

* Creating a Qt window inside Cocoa NSView

* re-using plugins (VST and AU) code

* Moving AU plugin sources to src/Plugins/AU

* Creating Jamtaba view in AU plugin

* Qt view and audio routing from DAW to AU plugin working

* Deleting plugin when deleted from daw

* Fixing dialogs

* Using small window

* Fixing resize

* Fixing cropped pan slider

* Fix the bug in QLineEDit focus when closing the host

* Multi input channels

* Fix APP_VERSION

* Updating mac installer to include AU plugin

+++ src/Common/MainController.h
@@ -173,7 +173,7 @@ public:
 
     void mixGroupedInputs(int groupIndex, Audio::SamplesBuffer &out);
 
-    virtual int getSampleRate() const = 0;
+    virtual float getSampleRate() const = 0;
 
     inline float getEncodingQuality() const
     {
